+++ src/Sentry.Unity.Editor.iOS/NativeOptions.cs
@@ -1,4 +1,6 @@
+using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 
 namespace Sentry.Unity.Editor.iOS
 {
@@ -8,6 +10,7 @@ namespace Sentry.Unity.Editor.iOS
 
         internal static string Generate(SentryUnityOptions options)
         {
+            var failedRequestStatusCodesArray = GetFailedRequestStatusCodesArray(options.FailedRequestStatusCodes);
             var nativeOptions = $@"#import <Foundation/Foundation.h>
 #import <Sentry/SentryOptions+HybridSDKs.h>
 #import <Sentry/PrivateSentrySDKOnly.h>
@@ -29,6 +32,8 @@ static SentryOptions* getSentryOptions()
         @""maxCacheItems"": @{options.MaxCacheItems},
         @""enableAutoSessionTracking"": @NO,
         @""enableAppHangTracking"": @NO,
+        @""enableCaptureFailedRequests"": @{ToObjCString(options.CaptureFailedRequests)},
+        @""failedRequestStatusCodes"" : @[{failedRequestStatusCodesArray}],
         @""sendDefaultPii"" : @{ToObjCString(options.SendDefaultPii)},
         @""attachScreenshot"" : @{ToObjCString(options.AttachScreenshot)},
         @""release"" : @""{options.Release}"",
@@ -67,6 +72,21 @@ static SentryOptions* getSentryOptions()
                 SentryLevel.Fatal => "fatal",
                 _ => "none"
             };
+        }
+
+        private static string GetFailedRequestStatusCodesArray(IList<HttpStatusCodeRange> httpStatusCodeRanges)
+        {
+            var codeRanges = string.Empty;
+            for (var i = 0; i < httpStatusCodeRanges.Count; i++)
+            {
+                codeRanges += $"[[SentryHttpStatusCodeRange alloc] initWithMin:{httpStatusCodeRanges[i].Start} max:{httpStatusCodeRanges[i].End}]";
+                if (i < httpStatusCodeRanges.Count - 1)
+                {
+                    codeRanges += ", ";
+                }
+            }
+
+            return codeRanges;
         }
     }
 }